refactor: rename SettingsButton
